Fix: '인증하지 않은 유저' 역할 추가, spec 구조 변경
